move helper functions from scraper into own globals file
