+++ scraper.js
@@ -1,39 +1,5 @@
 const api = require('./api/api.js')
-const { storeTransaction, getTransacton, lookupTransaction } = require('./keeper.js')
-
-// loop through the outputs of a tx, greedily returning the value of an output tx where n matches vOutIdx
-const getMatchingTransactionValue = async (txHash, voutIndex) => {
-  let tx = undefined
-  let voutArray = undefined
-
-  if (lookupTransaction(txHash)) {
-    voutArray = getTransacton(txHash)
-  } else {
-    tx = await api.getRawTransaction(txHash)
-    voutArray = JSON.parse(tx).vout
-    storeTransaction(txHash, voutArray)
-  }
-
-  for (let i = 0; i < voutArray.length; i++) {
-    if (voutArray[i].n === voutIndex) {
-      return voutArray[i].value
-    }
-  }
-}
-
-const sumOutputs = (voutArray) => {
-  return voutArray.reduce( (accumulator, currentValue) => accumulator + currentValue.value, 0)
-}
-
-const calculateFee = async (tx, outputTotal) => {
-  let inputTotal = 0
-
-  for (let i = 0; i < tx.vin.length; i++) {
-    inputTotal += await getMatchingTransactionValue(tx.vin[i].txid, tx.vin[i].vout)
-  }
-
-  return inputTotal - outputTotal
-}
+const globals = require('./globals.js')
 
 const scraper = async (blockHeight) => {
   let blockHash = await api.getBlockHashByHeight(blockHeight)
@@ -46,8 +12,8 @@ const scraper = async (blockHeight) => {
     let tx = await api.getRawTransaction(transactions[i])
     tx = JSON.parse(tx)
 
-    let txAmount = sumOutputs(tx.vout)
-    let fee = await calculateFee(tx, txAmount)
+    let txAmount = globals.getTransactionTotal(tx.vout)
+    let fee = await globals.calculateFee(tx, txAmount)
 
     blockTransactionData.push([blockHeight, txAmount, fee, tx.time, tx.txid])
   }
@@ -57,16 +23,6 @@ const scraper = async (blockHeight) => {
   return({ msg: 'blockDone', data: blockTransactionData, block: blockHeight })
 }
 
-const testTransaction = async (txHash) => {
-  const tx = await api.getRawTransaction(txHash)
-  const fee = await calculateFee(tx)
-
-  console.log(tx)
-  console.log(txHash)
-  console.log(fee)
-}
-
 module.exports = {
-  testTransaction,
   scraper
 }
